修改Count返回值

+++ src/main/java/com/shishuo/cms/service/AdminService.java
@@ -68,7 +68,7 @@ public class AdminService {
 	}
 
 	public int getAllListCount() {
-		return (int) adminDao.getAllListCount();
+		return adminDao.getAllListCount();
 	}
 
 	public PageVo<Admin> getAllListPage(int pageNum) {
